move our indexer and querier
